move canonical bridges back to checkbox
