--- packages/frontend/src/components/table/filters/BridgesTableFilters.stories.tsx
@@ -1,15 +1,0 @@
-import React from 'react'
-
-import { BridgesTableFilters as BridgesTableFiltersComponent } from './BridgesTableFilters'
-
-export default {
-  title: 'Components/Table/BridgesFilters',
-}
-
-export function BridgesFilters() {
-  return (
-    <div className="p-10">
-      <BridgesTableFiltersComponent />
-    </div>
-  )
-}
