+++ packages/frontend/src/pages/bridges-risk/BridgesRiskView.stories.tsx
@@ -194,15 +194,6 @@ export const Active: Story = {
   },
 }
 
-export const CanonicalBridges: Story = {
-  render: () => {
-    useEffect(() => {
-      click('.TabNavigationTab#canonical-bridges')
-    }, [])
-    return <Template />
-  },
-}
-
 export const Archived: Story = {
   render: () => {
     useEffect(() => {
