+++ packages/frontend/src/scripts/index.ts
@@ -2,6 +2,7 @@ import { configureCharts } from '../components/chart/configure'
 import { configureTvlActivity } from '../components/header/configureTvlActivity'
 import { configureDarkThemeToggle } from '../components/navbar/configureDarkThemeToggle'
 import { configureSidebarMenu } from '../components/navbar/configureSidebarMenu'
+import { configureCanonicalBridgesFilter } from './configureCanonicalBridgesFilter'
 import { configureDropdowns } from './configureDropdowns'
 import { configureExpandableContainer } from './configureExpandableContainer'
 import { configureHoverableDropdown } from './configureHoverableDropdown'
@@ -21,3 +22,4 @@ configureHoverableDropdown()
 configureProjectNavigation()
 configureExpandableContainer()
 configureTabNavigations()
+configureCanonicalBridgesFilter()
